# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "McAfee SIEM Alarm Editor API",
    "description": "REST API for managing SIEM customers, rules, and alarms",
    "version": "2.0.0",
    "contact": {
      "name": "McAfee SIEM Alarm Editor",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/customers/{id}/rules": {
      "get": {
        "tags": ["Rules"],
        "summary": "Get customer rules",
        "description": "Retrieve all rules for a specific customer with optional filtering",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Customer ID"
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Search term for rule name, ID, or SigID"
          },
          {
            "name": "severity_min",
            "in": "query",
            "schema": { "type": "integer", "minimum": 0, "maximum": 100 },
            "description": "Minimum severity filter"
          },
          {
            "name": "severity_max",
            "in": "query",
            "schema": { "type": "integer", "minimum": 0, "maximum": 100 },
            "description": "Maximum severity filter"
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "default": 1 },
            "description": "Page number for pagination"
          },
          {
            "name": "per_page",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of rules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "rules": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Rule" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/customers/{id}/rules/generate-alarms": {
      "post": {
        "tags": ["Rules"],
        "summary": "Generate alarms from rules",
        "description": "Generate alarms for selected rules",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Customer ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rule_ids": {
                    "type": "array",
                    "items": { "type": "integer" },
                    "description": "Array of rule IDs to generate alarms for"
                  }
                },
                "required": ["rule_ids"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alarms generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "generated_alarms": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Alarm" }
                    },
                    "generated_count": { "type": "integer" },
                    "errors": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Rule": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "customer_id": { "type": "integer" },
          "rule_id": { "type": "string", "example": "47-6000114" },
          "name": { "type": "string", "example": "Suspicious Login Activity" },
          "description": { "type": "string" },
          "severity": { "type": "integer", "minimum": 0, "maximum": 100 },
          "sig_id": { "type": "string", "example": "6000114" },
          "rule_type": { "type": "integer" },
          "revision": { "type": "integer" },
          "origin": { "type": "integer" },
          "action": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Alarm": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "customer_id": { "type": "integer" },
          "name": { "type": "string", "example": "High Severity Alert" },
          "min_version": { "type": "string", "example": "11.6.14" },
          "severity": { "type": "integer", "minimum": 0, "maximum": 100 },
          "match_field": { "type": "string", "example": "DSIDSigID" },
          "match_value": { "type": "string", "example": "47|6000114" },
          "condition_type": { "type": "integer" },
          "assignee_id": { "type": "integer" },
          "esc_assignee_id": { "type": "integer" },
          "note": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "per_page": { "type": "integer" },
          "total": { "type": "integer" },
          "pages": { "type": "integer" },
          "has_next": { "type": "boolean" },
          "has_prev": { "type": "boolean" }
        }
      },
...
    }
  },
...
}